style(vue-hoc): fix lint errors
